Add plotter functionality
